# apps/backend/app/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Awade API",
    "description": "AI-powered educator support platform for African teachers",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/curriculum/": {
      "post": {
        "tags": [
          "curriculum"
        ],
        "summary": "Create Curriculum",
        "description": "Create a new curriculum record.\nRequires admin authentication.",
        "operationId": "create_curriculum_api_curriculum__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CurriculumCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurriculumResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "curriculum"
        ],
        "summary": "Get Curriculums",
        "description": "Retrieve a list of curriculums, optionally filtered by country.\nRequires authentication.",
        "operationId": "get_curriculums_api_curriculum__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "country_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Country Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CurriculumResponse"
                  },
                  "title": "Response Get Curriculums Api Curriculum  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/curriculum/{curriculum_id}": {
      "get": {
        "tags": [
          "curriculum"
        ],
        "summary": "Get Curriculum",
        "description": "Retrieve a specific curriculum by ID.\nRequires authentication.",
        "operationId": "get_curriculum_api_curriculum__curriculum_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "curriculum_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Curriculum Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurriculumResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/curriculum-structures/": {
      "get": {
        "tags": [
          "curriculum-structures"
        ],
        "summary": "List Curriculum Structures",
        "description": "Retrieve a list of all curriculum structures.\nRequires authentication.",
        "operationId": "list_curriculum_structures_api_curriculum_structures__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CurriculumStructureResponse"
                  },
                  "type": "array",
                  "title": "Response List Curriculum Structures Api Curriculum Structures  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "curriculum-structures"
        ],
        "summary": "Create Curriculum Structure",
        "description": "Create a new curriculum structure record.\nRequires admin authentication.",
        "operationId": "create_curriculum_structure_api_curriculum_structures__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CurriculumStructureCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurriculumStructureResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/curriculum-structures/{structure_id}": {
      "get": {
        "tags": [
          "curriculum-structures"
        ],
        "summary": "Get Curriculum Structure",
        "description": "Get a specific curriculum structure by ID.\nRequires authentication.",
        "operationId": "get_curriculum_structure_api_curriculum_structures__structure_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "structure_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Structure Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurriculumStructureResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "curriculum-structures"
        ],
        "summary": "Update Curriculum Structure",
        "description": "Update a curriculum structure record.\nRequires admin authentication.",
        "operationId": "update_curriculum_structure_api_curriculum_structures__structure_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "structure_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Structure Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CurriculumStructureCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurriculumStructureResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "curriculum-structures"
        ],
        "summary": "Delete Curriculum Structure",
        "description": "Delete a curriculum structure record.\nRequires admin authentication.",
        "operationId": "delete_curriculum_structure_api_curriculum_structures__structure_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "structure_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Structure Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CurriculumCreate": {
        "properties": {
          "curricula_title": {
            "type": "string",
            "title": "Curricula Title"
          },
          "country_id": {
            "type": "integer",
            "title": "Country Id"
          }
        },
        "type": "object",
        "required": [
          "curricula_title",
          "country_id"
        ],
        "title": "CurriculumCreate",
        "description": "Schema for creating a new curriculum."
      },
      "CurriculumResponse": {
        "properties": {
          "curricula_title": {
            "type": "string",
            "title": "Curricula Title"
          },
          "country_id": {
            "type": "integer",
            "title": "Country Id"
          },
          "curricula_id": {
            "type": "integer",
            "title": "Curricula Id"
          }
        },
        "type": "object",
        "required": [
          "curricula_title",
          "country_id",
          "curricula_id"
        ],
        "title": "CurriculumResponse",
        "description": "Schema for curriculum response data."
      },
      "CurriculumStructureCreate": {
        "properties": {
          "curricula_id": {
            "type": "integer",
            "title": "Curricula Id"
          },
          "grade_level_id": {
            "type": "integer",
            "title": "Grade Level Id"
          },
          "subject_id": {
            "type": "integer",
            "title": "Subject Id"
          }
        },
        "type": "object",
        "required": [
          "curricula_id",
          "grade_level_id",
          "subject_id"
        ],
        "title": "CurriculumStructureCreate",
        "description": "Schema for creating a new curriculum structure."
      },
      "CurriculumStructureResponse": {
        "properties": {
          "curriculum_structure_id": {
            "type": "integer",
            "title": "Curriculum Structure Id"
          },
          "curricula_id": {
            "type": "integer",
            "title": "Curricula Id"
          },
          "grade_level_id": {
            "type": "integer",
            "title": "Grade Level Id"
          },
          "subject_id": {
            "type": "integer",
            "title": "Subject Id"
          }
        },
        "type": "object",
        "required": [
          "curriculum_structure_id",
          "curricula_id",
          "grade_level_id",
          "subject_id"
        ],
        "title": "CurriculumStructureResponse",
        "description": "Schema for curriculum structure response data."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}